+++ traiders/backend/api/models/__init__.py
@@ -1,4 +1,4 @@
 from .equipment import Equipment
 from .investment import ManualInvestment
 from .parity import Parity
-from .user import User
+from .users import User
