+++ traiders/backend/api/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .users import UserSerializer
