+++ traiders/backend/api/views/__init__.py
@@ -1,0 +1,1 @@
+from .users import UserViewSet
